REG-5204 Lagt til dyp sletting på øvrige artifacter

+++ dolly-app/src/main/java/no/nav/dolly/service/PersonService.java
@@ -5,14 +5,16 @@ import static java.util.Collections.singletonList;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException;
 
+import lombok.extern.slf4j.Slf4j;
+import no.nav.dolly.bestilling.ClientRegister;
 import no.nav.dolly.bestilling.tpsf.TpsfService;
 import no.nav.dolly.domain.jpa.Testgruppe;
 import no.nav.dolly.domain.jpa.Testident;
 
+@Slf4j
 @Service
 public class PersonService {
 
@@ -22,14 +24,14 @@ public class PersonService {
     @Autowired
     private TestgruppeService testgruppeService;
 
+    @Autowired
+    private List<ClientRegister> clientRegister;
+
     public void recyclePersoner(List<String> identer) {
 
-        try {
+        if (!identer.isEmpty()) {
             tpsfService.deletePersoner(identer);
-        } catch (HttpClientErrorException error) {
-            if (HttpStatus.NOT_FOUND.value() != error.getStatusCode().value()) {
-                throw error;
-            }
+            releaseArtifacts(identer);
         }
     }
 
@@ -42,5 +44,19 @@ public class PersonService {
 
         Testgruppe testgruppe = testgruppeService.fetchTestgruppeById(gruppeId);
         recyclePersoner(testgruppe.getTestidenter().stream().map(Testident::getIdent).collect(Collectors.toList()));
+    }
+
+    @Async
+    public void releaseArtifacts(List<String> identer) {
+
+        clientRegister.forEach(register -> {
+
+            try {
+                register.release(identer);
+
+            } catch (RuntimeException e) {
+                log.error("Feilet å slette fra register, {}", e.getMessage(), e);
+            }
+        });
     }
 }
